backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI and ML
openai==1.3.7
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# NLP and text processing
textblob==0.17.1
gensim==4.3.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP client and data fetching
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
